admin/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
rq
# For shared module if it was a package, but here we just copy code
